Split out some more transaction tests
